Moved value providers into dedicated folder for better maintenance.
