+++ src/main/java/com/mule/mulechain/vectors/internal/MuleChainVectorsModelParameters.java
@@ -1,8 +1,8 @@
 package com.mule.mulechain.vectors.internal;
 
+import com.mule.mulechain.vectors.internal.providers.MuleChainVectorsParameterEModelNameProvider;
 import org.mule.runtime.api.meta.ExpressionSupport;
 import org.mule.runtime.extension.api.annotation.Expression;
-import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.values.OfValues;
